requirements.txt
configparser>=3.5.0
requests>=2.21.0
pyyaml>=5.1.2
